requirements.txt
alabaster==0.7.12
anyio==4.4.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
arrow==1.3.0
asgiref==3.8.1
astropy==5.0.2
async-lru==2.0.4
attrs==23.2.0
autobahn==21.11.1
Automat==20.2.0
Babel==2.15.0
backcall==0.2.0
base58==1.0.3
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
brial==1.0.2
Brlapi==0.8.3
Brotli==1.0.9
cbor==1.0.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
cloud-init==23.1.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
corner==2.2.2
cov-core==1.15.0
coverage==6.2
crispy-bootstrap5==2024.10
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
cypari2==2.1.2
Cython==0.29.28
data-wizard==2.0.0
dbus-python==1.2.18
debugpy==1.8.2
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==5.1.2
django-cors-headers==4.6.0
django-crispy-forms==2.3
djangorestframework==3.15.2
docutils==0.17.1
ecdsa==0.18.0b1
emcee==3.1.6
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.2.1
fastjsonschema==2.20.0
filelock==3.6.0
flatbuffers===1.12.1-git20200711.33e2d80-dfsg1-0.6
fonttools==4.29.1
fpylll==0.5.2.dev0
fqdn==1.5.1
fs==2.4.12
future==0.18.2
gast==0.5.2
GDAL==3.4.1
git-cola==3.12.0
gmpy2==2.1.2
gunicorn==23.0.0
h11==0.14.0
htcondor==23.9.6
html-json-forms==1.1.1
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
hyperlink==21.0.0
hypothesis==6.36.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==1.1.1
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
isoduration==20.11.0
itertable==2.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.4.2
json5==0.9.25
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-console==6.4.0
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.27.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.0.1
matplotlib==3.7.2
matplotlib-inline==0.1.3
memory-allocator==0.1.2
mistune==3.0.2
mnemonic==0.19
mock==4.0.3
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
msgpack==1.0.3
natural-keys==2.1.1
nbclient==0.5.6
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==2.4
nose2==0.9.2
notebook==7.0.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
overrides==7.7.0
OWSLib==0.25.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.1
passlib==1.7.4
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkgconfig==1.5.5
platformdirs==2.6.2
plotly==5.4.0
pluggy==0.13.0
ply==3.11
pplpy==0.8.7
primecountpy==0.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.0
psycopg2==2.9.2
ptyprocess==0.7.0
py==1.10.0
py-ubjson==0.16.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydot==2.0.0
pyerfa==2.0.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylatexenc==2.10
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pypng==0.0.20
pyproj==3.3.0
PyQRCode==1.2.1
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
pytest==6.2.5
pytest-arraydiff==0.5.0
pytest-astropy==0.9.0
pytest-astropy-header==0.2.0
pytest-cov==3.0.0
pytest-doctestplus==0.11.2
pytest-filter-subpackage==0.1.1
pytest-mock==3.6.1
pytest-openfiles==0.5.0
pytest-remotedata==0.3.3
python-apt==2.4.0+ubuntu1
python-dateutil==2.9.0.post0
python-debian===0.1.43ubuntu1
python-json-logger==2.0.7
python-magic==0.4.24
python-snappy==0.5.3
pythran==0.10.0
PyTrie==0.4.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
qiskit==1.1.1
qiskit-aer==0.14.2
qiskit-algorithms==0.3.0
QScintilla==2.11.6
QtPy==2.0.0
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==3.3
rpds-py==0.18.1
rpy2==3.4.5
rustworkx==0.15.1
sagemath-standard==9.5
sagenb-export==3.2
scikit-learn==1.5.2
scipy==1.11.2
scour==0.38.2
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.1.0
sos==4.4
soupsieve==2.3.1
Sphinx==4.3.2
sqlparse==0.4.2
ssh-import-id==5.11
stevedore==5.2.0
symengine==0.11.0
sympy==1.9
systemd-python==234
tenacity==6.3.1
terminado==0.13.1
terminator==2.1.1
testpath==0.5.0
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
tox==3.21.4
traitlets==5.14.3
Twisted==22.1.0
txaio==21.2.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==2.1
u-msgpack-python==2.3.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.1.0
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.3.0
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.5
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==2.0.0
wsaccel==0.6.3
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
